Move guards to own folder
